Refactor: split oversized files into focused modules (250-line cap)
- Extract 8 tool registration modules into src/tools/
- Extract graph D1 operations into src/graph/ (namespaces, entities, relations, traversal)
- Extract OAuth utilities into src/oauth/ (error, sanitize, csrf, state, approval)
- Extract JWT verification to src/jwt.ts
- Extract response helpers to src/response-helpers.ts
- Consolidate 25 tools down to 13 action-based tools for token efficiency
- Update README with new tool surface (13 tools) and project structure
- Update AGENTS.md with file structure documentation
- All files now under 250 lines (largest: 197 lines)
- Bump agents 0.1.6 -> 0.6.0, MCP SDK 1.18.1 -> 1.27.1, zod 3.25.0 -> 4.3.6
- Add npm override to force single MCP SDK copy (agents pins exact version)
- Add react as devDep (agents peer dep, not bundled by wrangler)
