+++ src/oauth/approval.ts
@@ -1,160 +1,8 @@
-/**
- * OAuth utility functions with CSRF and state validation security.
- * Adapted from the Cloudflare Access MCP example.
- */
-import type { AuthRequest, ClientInfo } from "@cloudflare/workers-oauth-provider";
+/** Client approval cookies, HMAC signing, and approval dialog UI. */
+import type { ClientInfo } from "@cloudflare/workers-oauth-provider";
+import { sanitizeText, sanitizeUrl } from "./sanitize.js";
 
-/**
- * OAuth 2.1 compliant error class.
- */
-export class OAuthError extends Error {
-  constructor(
-    public code: string,
-    public description: string,
-    public statusCode = 400,
-  ) {
-    super(description);
-    this.name = "OAuthError";
-  }
-
-  toResponse(): Response {
-    return new Response(
-      JSON.stringify({
-        error: this.code,
-        error_description: this.description,
-      }),
-      {
-        status: this.statusCode,
-        headers: { "Content-Type": "application/json" },
-      },
-    );
-  }
-}
-
-export interface OAuthStateResult {
-  stateToken: string;
-}
-
-export interface ValidateStateResult {
-  oauthReqInfo: AuthRequest;
-  clearCookie: string;
-}
-
-export interface CSRFProtectionResult {
-  token: string;
-  setCookie: string;
-}
-
-export interface ValidateCSRFResult {
-  clearCookie: string;
-}
-
-function sanitizeText(text: string): string {
-  return text
-    .replace(/&/g, "&amp;")
-    .replace(/</g, "&lt;")
-    .replace(/>/g, "&gt;")
-    .replace(/"/g, "&quot;")
-    .replace(/'/g, "&#039;");
-}
-
-function sanitizeUrl(url: string): string {
-  const normalized = url.trim();
-  if (normalized.length === 0) return "";
-
-  for (let i = 0; i < normalized.length; i++) {
-    const code = normalized.charCodeAt(i);
-    if ((code >= 0x00 && code <= 0x1f) || (code >= 0x7f && code <= 0x9f)) {
-      return "";
-    }
-  }
-
-  let parsedUrl: URL;
-  try {
-    parsedUrl = new URL(normalized);
-  } catch {
-    return "";
-  }
-
-  const allowedSchemes = ["https", "http"];
-  const scheme = parsedUrl.protocol.slice(0, -1).toLowerCase();
-  if (!allowedSchemes.includes(scheme)) return "";
-
-  return normalized;
-}
-
-export function generateCSRFProtection(): CSRFProtectionResult {
-  const csrfCookieName = "__Host-CSRF_TOKEN";
-  const token = crypto.randomUUID();
-  const setCookie = `${csrfCookieName}=${token}; HttpOnly; Secure; Path=/; SameSite=Lax; Max-Age=600`;
-  return { token, setCookie };
-}
-
-export function validateCSRFToken(formData: FormData, request: Request): ValidateCSRFResult {
-  const csrfCookieName = "__Host-CSRF_TOKEN";
-  const tokenFromForm = formData.get("csrf_token");
-
-  if (!tokenFromForm || typeof tokenFromForm !== "string") {
-    throw new OAuthError("invalid_request", "Missing CSRF token in form data", 400);
-  }
-
-  const cookieHeader = request.headers.get("Cookie") || "";
-  const cookies = cookieHeader.split(";").map((c) => c.trim());
-  const csrfCookie = cookies.find((c) => c.startsWith(`${csrfCookieName}=`));
-  const tokenFromCookie = csrfCookie ? csrfCookie.substring(csrfCookieName.length + 1) : null;
-
-  if (!tokenFromCookie) {
-    throw new OAuthError("invalid_request", "Missing CSRF token cookie", 400);
-  }
-
-  if (tokenFromForm !== tokenFromCookie) {
-    throw new OAuthError("invalid_request", "CSRF token mismatch", 400);
-  }
-
-  const clearCookie = `${csrfCookieName}=; HttpOnly; Secure; Path=/; SameSite=Lax; Max-Age=0`;
-  return { clearCookie };
-}
-
-export async function createOAuthState(
-  oauthReqInfo: AuthRequest,
-  kv: KVNamespace,
-  stateTTL = 600,
-): Promise<OAuthStateResult> {
-  const stateToken = crypto.randomUUID();
-  await kv.put(`oauth:state:${stateToken}`, JSON.stringify(oauthReqInfo), {
-    expirationTtl: stateTTL,
-  });
-  return { stateToken };
-}
-
-export async function validateOAuthState(
-  request: Request,
-  kv: KVNamespace,
-): Promise<ValidateStateResult> {
-  const url = new URL(request.url);
-  const stateFromQuery = url.searchParams.get("state");
-
-  if (!stateFromQuery) {
-    throw new OAuthError("invalid_request", "Missing state parameter", 400);
-  }
-
-  const storedDataJson = await kv.get(`oauth:state:${stateFromQuery}`);
-  if (!storedDataJson) {
-    throw new OAuthError("invalid_request", "Invalid or expired state", 400);
-  }
-
-  let oauthReqInfo: AuthRequest;
-  try {
-    oauthReqInfo = JSON.parse(storedDataJson) as AuthRequest;
-  } catch {
-    throw new OAuthError("server_error", "Invalid state data", 500);
-  }
-
-  // Delete state from KV (one-time use)
-  await kv.delete(`oauth:state:${stateFromQuery}`);
-
-  return { oauthReqInfo, clearCookie: "" };
-}
+// --- Client approval cookie management ---
 
 export async function isClientApproved(
   request: Request,
@@ -183,76 +31,7 @@ export async function addApprovedClient(
   return `${approvedClientsCookieName}=${cookieValue}; HttpOnly; Secure; Path=/; SameSite=Lax; Max-Age=${THIRTY_DAYS_IN_SECONDS}`;
 }
 
-export function getUpstreamAuthorizeUrl(params: {
-  upstream_url: string;
-  client_id: string;
-  redirect_uri: string;
-  scope: string;
-  state: string;
-}): string {
-  const url = new URL(params.upstream_url);
-  url.searchParams.set("client_id", params.client_id);
-  url.searchParams.set("redirect_uri", params.redirect_uri);
-  url.searchParams.set("response_type", "code");
-  url.searchParams.set("scope", params.scope);
-  url.searchParams.set("state", params.state);
-  return url.toString();
-}
-
-export async function fetchUpstreamAuthToken(params: {
-  upstream_url: string;
-  client_id: string;
-  client_secret: string;
-  code?: string;
-  redirect_uri: string;
-}): Promise<[string, string, null] | [null, null, Response]> {
-  if (!params.code) {
-    return [null, null, new Response("Missing authorization code", { status: 400 })];
-  }
-
-  const data = new URLSearchParams({
-    client_id: params.client_id,
-    client_secret: params.client_secret,
-    code: params.code,
-    grant_type: "authorization_code",
-    redirect_uri: params.redirect_uri,
-  });
-
-  const response = await fetch(params.upstream_url, {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/x-www-form-urlencoded",
-      Accept: "application/json",
-    },
-    body: data.toString(),
-  });
-
-  if (!response.ok) {
-    const errorText = await response.text();
-    return [
-      null,
-      null,
-      new Response(`Failed to exchange code for token: ${errorText}`, {
-        status: response.status,
-      }),
-    ];
-  }
-
-  const body = (await response.json()) as Record<string, unknown>;
-
-  const accessToken = body.access_token as string;
-  if (!accessToken) {
-    return [null, null, new Response("Missing access token", { status: 400 })];
-  }
-
-  const idToken = body.id_token as string;
-  if (!idToken) {
-    return [null, null, new Response("Missing id token", { status: 400 })];
-  }
-  return [accessToken, idToken, null];
-}
-
-// --- Approval dialog ---
+// --- Approval dialog UI ---
 
 export interface ApprovalDialogOptions {
   client: ClientInfo | null;
@@ -339,7 +118,7 @@ export function renderApprovalDialog(request: Request, options: ApprovalDialogOp
   });
 }
 
-// --- Internal helpers ---
+// --- Internal crypto helpers ---
 
 async function getApprovedClientsFromCookie(
   request: Request,
