+++ src/oauth/csrf.ts
@@ -1,0 +1,43 @@
+/** Double-submit cookie CSRF protection. */
+import { OAuthError } from "./error.js";
+
+export interface CSRFProtectionResult {
+  token: string;
+  setCookie: string;
+}
+
+export interface ValidateCSRFResult {
+  clearCookie: string;
+}
+
+export function generateCSRFProtection(): CSRFProtectionResult {
+  const csrfCookieName = "__Host-CSRF_TOKEN";
+  const token = crypto.randomUUID();
+  const setCookie = `${csrfCookieName}=${token}; HttpOnly; Secure; Path=/; SameSite=Lax; Max-Age=600`;
+  return { token, setCookie };
+}
+
+export function validateCSRFToken(formData: FormData, request: Request): ValidateCSRFResult {
+  const csrfCookieName = "__Host-CSRF_TOKEN";
+  const tokenFromForm = formData.get("csrf_token");
+
+  if (!tokenFromForm || typeof tokenFromForm !== "string") {
+    throw new OAuthError("invalid_request", "Missing CSRF token in form data", 400);
+  }
+
+  const cookieHeader = request.headers.get("Cookie") || "";
+  const cookies = cookieHeader.split(";").map((c) => c.trim());
+  const csrfCookie = cookies.find((c) => c.startsWith(`${csrfCookieName}=`));
+  const tokenFromCookie = csrfCookie ? csrfCookie.substring(csrfCookieName.length + 1) : null;
+
+  if (!tokenFromCookie) {
+    throw new OAuthError("invalid_request", "Missing CSRF token cookie", 400);
+  }
+
+  if (tokenFromForm !== tokenFromCookie) {
+    throw new OAuthError("invalid_request", "CSRF token mismatch", 400);
+  }
+
+  const clearCookie = `${csrfCookieName}=; HttpOnly; Secure; Path=/; SameSite=Lax; Max-Age=0`;
+  return { clearCookie };
+}
