+++ src/oauth/error.ts
@@ -1,0 +1,24 @@
+/** OAuth 2.1 compliant error class. */
+export class OAuthError extends Error {
+  constructor(
+    public code: string,
+    public description: string,
+    public statusCode = 400,
+  ) {
+    super(description);
+    this.name = "OAuthError";
+  }
+
+  toResponse(): Response {
+    return new Response(
+      JSON.stringify({
+        error: this.code,
+        error_description: this.description,
+      }),
+      {
+        status: this.statusCode,
+        headers: { "Content-Type": "application/json" },
+      },
+    );
+  }
+}
