+++ src/oauth/sanitize.ts
@@ -1,0 +1,35 @@
+/** Input sanitization for the OAuth approval dialog. */
+
+export function sanitizeText(text: string): string {
+  return text
+    .replace(/&/g, "&amp;")
+    .replace(/</g, "&lt;")
+    .replace(/>/g, "&gt;")
+    .replace(/"/g, "&quot;")
+    .replace(/'/g, "&#039;");
+}
+
+export function sanitizeUrl(url: string): string {
+  const normalized = url.trim();
+  if (normalized.length === 0) return "";
+
+  for (let i = 0; i < normalized.length; i++) {
+    const code = normalized.charCodeAt(i);
+    if ((code >= 0x00 && code <= 0x1f) || (code >= 0x7f && code <= 0x9f)) {
+      return "";
+    }
+  }
+
+  let parsedUrl: URL;
+  try {
+    parsedUrl = new URL(normalized);
+  } catch {
+    return "";
+  }
+
+  const allowedSchemes = ["https", "http"];
+  const scheme = parsedUrl.protocol.slice(0, -1).toLowerCase();
+  if (!allowedSchemes.includes(scheme)) return "";
+
+  return normalized;
+}
